backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
typing_extensions==4.15.0
django-phonenumber-field[phonenumberslite]==8.0.0
phonenumberslite==8.13.44
psycopg2-binary>=2.9
djangorestframework==3.16.1